agent-recruiter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-recruiter"
version = "0.1.0"
description = "Autonomous talent sourcing pipeline powered by multi-agent AI"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Daniel López Orta", email = "daniellopezorta39@gmail.com" }]
keywords = [
    "agent", "recruiting", "talent-sourcing", "resume-screening",
    "multi-agent", "llm", "hiring", "hr-tech",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "openai>=1.20",
    "tenacity>=8.3",
    "aiohttp>=3.9",
    "pymupdf>=1.24",
    "python-docx>=1.0",
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-instrumentation-openai>=0.23b0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4"]

[project.scripts]
agent-recruiter = "agent_recruiter.cli:main"

[project.urls]
Homepage = "https://github.com/daniellopez882/agent-recruiter"
Repository = "https://github.com/daniellopez882/agent-recruiter"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
